third_party/WebKit/LayoutTests/fast/css/fontface-methods-expected.txt

Tests load() and ready() methods of FontFace.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS face1.status is "unloaded"
PASS face1.status is "loading"
PASS face1.status is "loaded"
PASS face2.status is "unloaded"
PASS face2.status is "loading"
PASS face2.status is "error"
PASS successfullyParsed is true

TEST COMPLETE
